==> Analysis_Using_Python/02_Data_Cleaning.ipynb <==
# Data Cleaning

---
Import Libraries:

import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)
pd.set_option("display.float_format", "{:.2f}".format)

Load Raw Data:

file_path = ('../Data/01_Raw_Data/Phoenix_Global_Sales_Raw_Data.csv')
df = pd.read_csv(file_path)

df.head()

Checking Missing Values:

df.isnull().sum()

- Handling Missing Values:

# Numeric columns -> Fill With Median
numeric_cols = df.select_dtypes(include=["float64", "int64"]).columns

for col in numeric_cols:
    df[col] = df[col].fillna(df[col].median())

# Categorical Columns -> Fill with Mode
categorical_cols = df.select_dtypes(include=["object"]).columns

for col in categorical_cols:
    df[col] = df[col].fillna(df[col].mode()[0])

Remove Duplicates:

before = df.shape[0]
df.drop_duplicates(inplace = True)
after = df.shape[0]

print("Duplicates removed:", before - after)

Convert Data Types:

df['Date'] = pd.to_datetime(df['Date'])
df['Year'] = df['Year'].astype(int)
df['Units_Sold'] = df['Units_Sold'].astype(int)
df['Delivery_Days'] = df['Delivery_Days'].astype(int)

Standardize Text Columns:

text_cols = ['Country', 'Region', 'Product_Type', 'Sales_Channel', 'Customer_Type']

for col in text_cols:
    df[col] = df[col].str.strip().str.title()

Validate Revenue, Cost, Profit:

- Remove unrealistic values:-

# Revenue and Cost should not be Negative:-

df = df[(df['Revenue'] >= 0) & (df['Cost'] >= 0)]

- Recalculate Profit (Business Accuracy):-

df['Profit'] = df['Revenue'] - df['Cost']

Handle Outliers (IQR Method):

def remove_outliers(col):
    Q1 = df[col].quantile(0.25)
    Q3 = df[col].quantile(0.75)
    IQR = Q3 - Q1
    lower = Q1 - 1.5 * IQR
    upper = Q3 + 1.5 * IQR
    return df[(df[col] >= lower) & (df[col] <= upper)]

for col in ['Revenue', 'Cost', 'Profit', 'Units_Sold', 'Delivery_Days']:
    df = remove_outliers(col)


Validate Year with Date:

df['Extracted_Year'] = df['Date'].dt.year
df = df[df['Year'] == df['Extracted_Year']]
df.drop(columns=['Extracted_Year'], inplace=True)

Final Data:

df.info()
df.isnull().sum()
df.describe()

Export Cleaned Dataset:

clean_file_path = ("../Data/02_Clean_Data/Phoenix_Global_Sales_Clean_Data.csv")
df.to_csv(clean_file_path, index = False)

print("Cleaned data exported successfully!")

---

# Observations:-
1. Missing values handled using median/mode.
2. Duplicate records removed.
3. Data types standardized.
4. Negative revenue and cost removed.
5. Profit recalculated for accuracy.
6. Outliers handled using IQR method.
7. Year validated with date field.
8. Clean dataset exported for further processing.

---
